Move iconURLFor/getRandomBackgroundURL to @cardstack/runtime-common
Extract realm icon and background URL functions into a shared module
at runtime-common/realm-display-defaults.ts. Host app and software-factory
now import from the same source, eliminating three duplicate copies.
